+++ corerl/data_pipeline/all_the_time.py
@@ -6,7 +6,7 @@ import numpy as np
 import pandas as pd
 
 from corerl.component.network.utils import tensor
-from corerl.configs.config import MISSING, config, interpolate
+from corerl.configs.config import config, interpolate
 from corerl.data_pipeline.datatypes import PipelineFrame, StageCode, Step, Transition
 from corerl.data_pipeline.tag_config import TagConfig
 
@@ -16,7 +16,7 @@ class AllTheTimeTCConfig:
     name: str = "all-the-time"
     gamma: float = interpolate('${experiment.gamma}')
     min_n_step: int = 1
-    max_n_step: int = MISSING
+    max_n_step: int | None = None
 
 
 @dataclass(init=False)
@@ -75,6 +75,7 @@ class AllTheTimeTC:
 
         self.gamma = cfg.gamma
         self.min_n_step = cfg.min_n_step
+        assert cfg.max_n_step is not None
         self.max_n_step = cfg.max_n_step
         assert self.min_n_step > 0
         assert self.max_n_step >= self.min_n_step