+++ corerl/interaction/deployment_interaction.py
@@ -20,7 +20,7 @@ from corerl.messages.heartbeat import Heartbeat, HeartbeatConfig
 from corerl.state import AppState
 from corerl.utils.time import clock_generator, split_into_chunks, wait_for_timestamp
 
-logger = logging.getLogger(__file__)
+logger = logging.getLogger(__name__)
 
 @config()
 class DepInteractionConfig:
@@ -28,6 +28,7 @@ class DepInteractionConfig:
     historical_batch_size: int = 10000
     checkpoint_path: Path = Path('outputs/checkpoints')
     heartbeat: HeartbeatConfig = Field(default_factory=HeartbeatConfig)
+    warmup_period: timedelta | None = None
 
 
 class DeploymentInteraction(Interaction):
@@ -75,6 +76,8 @@ class DeploymentInteraction(Interaction):
             width=self.obs_period * cfg.historical_batch_size
         )
 
+        # warmup pipeline
+        self.warmup_pipeline()
         # checkpointing state
         self._last_checkpoint = datetime.now(UTC)
         self.restore_checkpoint()
@@ -104,7 +107,7 @@ class DeploymentInteraction(Interaction):
         self._app_state.agent_step += 1
 
     def step_event(self, event: Event):
-        logger.debug(f"Received step_event: {event}")
+        logger.debug(f"Interaction received Event: {event}")
         self._heartbeat.healthcheck()
         match event.type:
             case EventType.step:
@@ -125,11 +128,14 @@ class DeploymentInteraction(Interaction):
             case EventType.step_emit_action:
                 s = self._get_latest_state()
                 if s is not None:
+                    logger.info("Querying agent policy for new action")
                     a = self._agent.get_action(s)
                     a_df = self._pipeline.action_constructor.assign_action_names(a)
                     a_df = self._pipeline.preprocessor.inverse(a_df)
                     self._env.emit_action(a_df)
                     self._last_action = a_df
+                else:
+                    logger.info("New action requested, but no valid state was available")
 
             case EventType.ping_setpoints:
                 if self._last_action is not None:
@@ -137,6 +143,19 @@ class DeploymentInteraction(Interaction):
 
             case _:
                 logger.warning(f"Unexpected step_event: {event}")
+
+    def warmup_pipeline(self):
+        if self._cfg.warmup_period is None:
+            return
+
+        warmup_end = datetime.now(UTC)
+        warmup_obs = self._env.data_reader.batch_aggregated_read(
+            names=self._env.tag_names,
+            start_time=warmup_end - self._cfg.warmup_period,
+            end_time=warmup_end,
+            bucket_width=self.obs_period
+        )
+        self._pipeline(warmup_obs, caller_code=CallerCode.ONLINE)
 
     def load_historical_chunk(self):
         try: